refactor: getVideoHTML로 video html 렌더링 함수화
