+++ src/js/index.ts
@@ -1,9 +1,9 @@
 import { initController } from './controller/index.js';
-import { renderView } from './view/index.js'
+import { renderView } from './view/index.js';
 
 const app = () => {
   renderView();
   initController();
-}
+};
 
 app();
